removed client from main
